# backend/docs/swagger.yaml
definitions:
  errs.Code:
    enum:
    - 100001
    - 100002
    - 100003
    - 100004
    type: integer
    x-enum-varnames:
    - ErrUnknown
    - ErrInvalidRequestParameters
    - ErrServerPanic
    - ErrServerExiting
...
  protocol.OpenWebApplyReq:
    properties:
      accountId:
        description: 凭证 ID
        example: test_account
        maxLength: 64
        type: string
      authorities:
        description: 授权项
        example:
        - 1
        - 2
        - 3
        items:
          type: integer
        type: array
        uniqueItems: true
      frequency:
        description: 请求频率
        example: 100
        maximum: 500
        minimum: 10
        type: integer
      requestIp:
        description: 请求源，逗号分割
        example: '*,192.168.0.1,127.0.0.1-127.0.0.128'
        maxLength: 256
        type: string
    required:
    - accountId
    - frequency
    - requestIp
    type: object
  protocol.OpenWebApplyRsp:
    properties:
      password:
        description: 凭证密码
        type: string
    type: object
  protocol.OpenWebGetInformationRsp:
    properties:
      accountId:
        description: 凭证 ID
        example: test_account
        type: string
      appId:
        description: 应用 ID
        example: 4ef83c03e2ce4f1f94c11168d1acd087
        type: string
      authorities:
        description: 授权项
        example:
        - 1
        - 2
        - 3
        items:
          type: integer
        type: array
      createdTime:
        description: 创建时间
        example: "2020-01-01 01:01:01"
        type: string
      creator:
        description: 创建人
        example: zhangsan
        type: string
      expiration:
        description: 过期时间
        example: "2020-01-01 01:01:01"
        type: string
      frequency:
        description: 请求频率
        example: 100
        type: integer
      requestIp:
        description: 请求源
        example: '*,192.168.0.1,127.0.0.1-127.0.0.128'
        type: string
    type: object
  protocol.OpenWebListItem:
    properties:
      accountId:
        description: 凭证 ID
        example: test_account
        type: string
      authorities:
        description: 授权项
        example:
        - 1
        - 2
        - 3
        items:
          type: integer
        type: array
      createdTime:
        description: 创建时间
        example: "2020-01-01 01:01:01"
        type: string
      creator:
        description: 创建人
        example: zhangsan
        type: string
      expiration:
        description: 过期时间
        example: "2020-01-01 01:01:01"
        type: string
      frequency:
        description: 请求频率
        example: 100
        type: integer
      requestIp:
        description: 请求源
        example: '*,192.168.0.1,127.0.0.1-127.0.0.128'
        type: string
    type: object
  protocol.OpenWebListRsp:
    properties:
      count:
        description: 总数
        example: 10
        type: integer
      list:
        description: 每项信息
        items:
          $ref: '#/definitions/protocol.OpenWebListItem'
        type: array
    type: object
  protocol.OpenWebResetRsp:
    properties:
      password:
        description: 凭证密码
        type: string
    type: object
  protocol.OpenWebUpdateReq:
    properties:
      accountId:
        description: 凭证 ID
        example: test_account
        maxLength: 64
        type: string
      authorities:
        description: 授权项
        example:
        - 1
        - 2
        - 3
        items:
          type: integer
        type: array
        uniqueItems: true
      frequency:
        description: 请求频率
        example: 100
        maximum: 500
        minimum: 10
        type: integer
      requestIp:
        description: 请求源
        example: '*,192.168.0.1,127.0.0.1-127.0.0.128'
        maxLength: 256
        type: string
    required:
    - accountId
    - frequency
    - requestIp
    type: object
...
  util.Response-any:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/errs.Code'
        description: 响应码，大于 0 表示错误
        example: 0
      data:
        description: 响应数据
      message:
        description: 提示语
        example: 失败时的提示语
        type: string
    type: object
...
  util.Response-protocol_OpenWebApplyRsp:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/errs.Code'
        description: 响应码，大于 0 表示错误
        example: 0
      data:
        allOf:
        - $ref: '#/definitions/protocol.OpenWebApplyRsp'
        description: 响应数据
      message:
        description: 提示语
        example: 失败时的提示语
        type: string
    type: object
  util.Response-protocol_OpenWebGetInformationRsp:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/errs.Code'
        description: 响应码，大于 0 表示错误
        example: 0
      data:
        allOf:
        - $ref: '#/definitions/protocol.OpenWebGetInformationRsp'
        description: 响应数据
      message:
        description: 提示语
        example: 失败时的提示语
        type: string
    type: object
  util.Response-protocol_OpenWebListRsp:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/errs.Code'
        description: 响应码，大于 0 表示错误
        example: 0
      data:
        allOf:
        - $ref: '#/definitions/protocol.OpenWebListRsp'
        description: 响应数据
      message:
        description: 提示语
        example: 失败时的提示语
        type: string
    type: object
  util.Response-protocol_OpenWebResetRsp:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/errs.Code'
        description: 响应码，大于 0 表示错误
        example: 0
      data:
        allOf:
        - $ref: '#/definitions/protocol.OpenWebResetRsp'
        description: 响应数据
      message:
        description: 提示语
        example: 失败时的提示语
        type: string
    type: object
...
host: 127.0.0.1:8090
info:
  contact:
    email: ivfzhou@126.com
    name: ivfzhou
  license:
    name: MulanPSL-2.0 license
    url: http://license.coscl.org.cn/MulanPSL2
  title: CSMS API
  version: "1.0"
paths:
...
  /web/open/apply/{appId}:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求日期
        example: Mon, 02 Jan 2006 15:04:05 GMT
        in: header
        name: Date
        required: true
        type: string
      - description: 会话凭据
        example: csms_user=; csms_seesion=
        in: header
        name: Cookie
        required: true
        type: string
      - description: 应用 ID
        example: 4ef83c03e2ce4f1f94c11168d1acd087
        in: path
        name: appId
        required: true
        type: string
      - description: 请求体
        in: body
        name: _
        required: true
        schema:
          $ref: '#/definitions/protocol.OpenWebApplyReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response-protocol_OpenWebApplyRsp'
      summary: 申请
      tags:
      - Open-WebAPI
  /web/open/getInformation/{appId}:
    get:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: 请求日期
        example: Mon, 02 Jan 2006 15:04:05 GMT
        in: header
        name: Date
        required: true
        type: string
      - description: 会话凭据
        example: csms_user=; csms_seesion=
        in: header
        name: Cookie
        required: true
        type: string
      - description: 应用 ID
        example: 4ef83c03e2ce4f1f94c11168d1acd087
        in: path
        name: appId
        required: true
        type: string
      - description: 凭证 ID
        example: test_account
        in: query
        maxLength: 64
        name: accountId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response-protocol_OpenWebGetInformationRsp'
      summary: 获取请求凭证信息
      tags:
      - Open-WebAPI
  /web/open/list/{appId}:
    get:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: 请求日期
        example: Mon, 02 Jan 2006 15:04:05 GMT
        in: header
        name: Date
        required: true
        type: string
      - description: 会话凭据
        example: csms_user=; csms_seesion=
        in: header
        name: Cookie
        required: true
        type: string
      - description: 应用 ID
        example: 4ef83c03e2ce4f1f94c11168d1acd087
        in: path
        name: appId
        required: true
        type: string
      - description: 页数
        example: 1
        in: query
        name: pageNumber
        type: integer
      - description: 每页条数
        example: 10
        in: query
        maximum: 100
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response-protocol_OpenWebListRsp'
      summary: 获取请求凭证信息列表
      tags:
      - Open-WebAPI
  /web/open/remove/{appId}:
    delete:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: 请求日期
        example: Mon, 02 Jan 2006 15:04:05 GMT
        in: header
        name: Date
        required: true
        type: string
      - description: 会话凭据
        example: csms_user=; csms_seesion=
        in: header
        name: Cookie
        required: true
        type: string
      - description: 应用 ID
        example: 4ef83c03e2ce4f1f94c11168d1acd087
        in: path
        name: appId
        required: true
        type: string
      - description: 凭证 ID
        example: test_account
        in: query
        maxLength: 64
        name: accountId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response-any'
      summary: 删除
      tags:
      - Open-WebAPI
  /web/open/renewal/{appId}:
    get:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: 请求日期
        example: Mon, 02 Jan 2006 15:04:05 GMT
        in: header
        name: Date
        required: true
        type: string
      - description: 会话凭据
        example: csms_user=; csms_seesion=
        in: header
        name: Cookie
        required: true
        type: string
      - description: 应用 ID
        example: 4ef83c03e2ce4f1f94c11168d1acd087
        in: path
        name: appId
        required: true
        type: string
      - description: 凭证 ID
        example: test_account
        in: query
        maxLength: 64
        name: accountId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response-any'
      summary: 续期
      tags:
      - Open-WebAPI
  /web/open/reset/{appId}:
    get:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: 请求日期
        example: Mon, 02 Jan 2006 15:04:05 GMT
        in: header
        name: Date
        required: true
        type: string
      - description: 会话凭据
        example: csms_user=; csms_seesion=
        in: header
        name: Cookie
        required: true
        type: string
      - description: 应用 ID
        example: 4ef83c03e2ce4f1f94c11168d1acd087
        in: path
        name: appId
        required: true
        type: string
      - description: 凭证 ID
        example: test_account
        in: query
        maxLength: 64
        name: accountId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response-protocol_OpenWebResetRsp'
      summary: 重置密钥
      tags:
      - Open-WebAPI
  /web/open/update/{appId}:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求日期
        example: Mon, 02 Jan 2006 15:04:05 GMT
        in: header
        name: Date
        required: true
        type: string
      - description: 会话凭据
        example: csms_user=; csms_seesion=
        in: header
        name: Cookie
        required: true
        type: string
      - description: 应用 ID
        example: 4ef83c03e2ce4f1f94c11168d1acd087
        in: path
        name: appId
        required: true
        type: string
      - description: 请求体
        in: body
        name: _
        required: true
        schema:
          $ref: '#/definitions/protocol.OpenWebUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response-any'
      summary: 修改
      tags:
      - Open-WebAPI
...
schemes:
- http
swagger: "2.0"